# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Roadmap Accountability System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface AuthUser {
  id: string;
  username: string;
  firstName?: string;
  lastName?: string;
  profileImage?: string | null;
  isAuthenticated: boolean;
}

export interface DailyCheckin {
  id: number;
  userId: string;
  date: string;
  hasUni: boolean;
  hasOffice: boolean;
  tasks: string;
  availableHours: number;
  pomodoroSessions: number;
  hoursLogged: number;
  aiSchedule?: string | null;
  createdAt: string;
}

export interface CreateCheckinBody {
  hasUni: boolean;
  hasOffice: boolean;
  tasks: string;
  availableHours: number;
}

export interface PomodoroSession {
  id: number;
  userId: string;
  date: string;
  workDuration: number;
  breakDuration: number;
  completedAt: string;
}

export interface CreatePomodoroSessionBody {
  workDuration: number;
  breakDuration: number;
}

export interface EpisodeProgress {
  id: number;
  userId: string;
  episodeId: string;
  episodeTitle: string;
  phase: string;
  completedAt: string;
}

export interface LogEpisodeBody {
  episodeId: string;
  episodeTitle: string;
  phase: string;
}

export interface DailyHoursLog {
  date: string;
  hours: number;
  sessions: number;
}

export interface ProgressStats {
  totalHours: number;
  currentStreak: number;
  longestStreak: number;
  totalEpisodes: number;
  totalPomodoros: number;
  hoursThisWeek: number;
  hoursThisMonth: number;
  lastCheckIn?: string | null;
}

export interface GenerateScheduleBody {
  hasUni: boolean;
  hasOffice: boolean;
  tasks: string;
  availableHours: number;
  currentPhase: string;
}

export interface AIScheduleResponse {
  schedule: string;
  tasks: string[];
  pomodoroRecommendation: string;
}

export interface AIWeeklyReportResponse {
  report: string;
  hoursLogged: number;
  episodesCompleted: number;
  nextWeekGoal: string;
  motivationalMessage: string;
}

export interface UserProfile {
  id: string;
  username: string;
  firstName?: string;
  currentPhase: string;
  totalHoursLogged: number;
  currentStreak: number;
  longestStreak: number;
  lastCheckIn?: string | null;
  achievements: string[];
}

export interface UpdateUserProfileBody {
  currentPhase?: string;
  firstName?: string;
}

export interface AuthUserDetails {
  id: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUserDetails | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export interface LogoutSuccess {
  success: boolean;
}

export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type GetPomodoroSessionsParams = {
  limit?: number;
};

export type GetHoursLogParams = {
  days?: number;
};