# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Roadmap Accountability System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /checkin/today:
    get:
      operationId: getTodayCheckin
      tags: [checkin]
      summary: Get today's check-in if it exists
      responses:
        "200":
          description: Today's check-in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyCheckin"
        "404":
          description: No check-in today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /checkin:
    post:
      operationId: createCheckin
      tags: [checkin]
      summary: Create or update today's check-in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckinBody"
      responses:
        "200":
          description: Check-in created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyCheckin"
  /pomodoro/sessions:
    get:
      operationId: getPomodoroSessions
      tags: [pomodoro]
      summary: Get recent pomodoro sessions
      parameters:
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PomodoroSession"
    post:
      operationId: createPomodoroSession
      tags: [pomodoro]
      summary: Log a completed pomodoro session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePomodoroSessionBody"
      responses:
        "200":
          description: Session logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PomodoroSession"
  /progress/episodes:
    get:
      operationId: getEpisodeProgress
      tags: [progress]
      summary: Get all episode completion records
      responses:
        "200":
          description: Episode progress list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EpisodeProgress"
    post:
      operationId: logEpisodeCompletion
      tags: [progress]
      summary: Mark an episode as completed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogEpisodeBody"
      responses:
        "200":
          description: Episode logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EpisodeProgress"
  /progress/episodes/{episodeId}:
    delete:
      operationId: removeEpisodeCompletion
      tags: [progress]
      summary: Unmark an episode as completed
      parameters:
        - name: episodeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Episode removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /progress/hours:
    get:
      operationId: getHoursLog
      tags: [progress]
      summary: Get hours logged (heatmap data)
      parameters:
        - name: days
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Hours log
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyHoursLog"
...
  /ai/schedule:
    post:
      operationId: generateSchedule
      tags: [ai]
      summary: Generate AI daily schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateScheduleBody"
      responses:
        "200":
          description: AI generated schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIScheduleResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...
    DailyCheckin:
      type: object
      properties:
        id:
          type: number
        userId:
          type: string
        date:
          type: string
        hasUni:
          type: boolean
        hasOffice:
          type: boolean
        tasks:
          type: string
        availableHours:
          type: number
        pomodoroSessions:
          type: number
        hoursLogged:
          type: number
        aiSchedule:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, userId, date, hasUni, hasOffice, tasks, availableHours, pomodoroSessions, hoursLogged, createdAt]
    CreateCheckinBody:
      type: object
      properties:
        hasUni:
          type: boolean
        hasOffice:
          type: boolean
        tasks:
          type: string
        availableHours:
          type: number
      required: [hasUni, hasOffice, tasks, availableHours]
    PomodoroSession:
      type: object
      properties:
        id:
          type: number
        userId:
          type: string
        date:
          type: string
        workDuration:
          type: number
        breakDuration:
          type: number
        completedAt:
          type: string
      required: [id, userId, date, workDuration, breakDuration, completedAt]
    CreatePomodoroSessionBody:
      type: object
      properties:
        workDuration:
          type: number
        breakDuration:
          type: number
      required: [workDuration, breakDuration]
    EpisodeProgress:
      type: object
      properties:
        id:
          type: number
        userId:
          type: string
        episodeId:
          type: string
        episodeTitle:
          type: string
        phase:
          type: string
        completedAt:
          type: string
      required: [id, userId, episodeId, episodeTitle, phase, completedAt]
    LogEpisodeBody:
      type: object
      properties:
        episodeId:
          type: string
        episodeTitle:
          type: string
        phase:
          type: string
      required: [episodeId, episodeTitle, phase]
    DailyHoursLog:
      type: object
      properties:
        date:
          type: string
        hours:
          type: number
        sessions:
          type: number
      required: [date, hours, sessions]
...
    GenerateScheduleBody:
      type: object
      properties:
        hasUni:
          type: boolean
        hasOffice:
          type: boolean
        tasks:
          type: string
        availableHours:
          type: number
        currentPhase:
          type: string
      required: [hasUni, hasOffice, tasks, availableHours, currentPhase]
    AIScheduleResponse:
      type: object
      properties:
        schedule:
          type: string
        tasks:
          type: array
          items:
            type: string
        pomodoroRecommendation:
          type: string
      required: [schedule, tasks, pomodoroRecommendation]
...