# lib/api-zod/src/generated/types/createCheckinBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Roadmap Accountability System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCheckinBody {
  hasUni: boolean;
  hasOffice: boolean;
  tasks: string;
  availableHours: number;
}