# lib/api-zod/src/generated/types/dailyCheckin.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Roadmap Accountability System API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyCheckin {
  id: number;
  userId: string;
  date: string;
  hasUni: boolean;
  hasOffice: boolean;
  tasks: string;
  availableHours: number;
  pomodoroSessions: number;
  hoursLogged: number;
  aiSchedule?: string | null;
  createdAt: string;
}